Version 0.2.0.0: Added Satchel UI, Wallrun "Animation", added assets for level design.

+++ Source/BotBlastFinal/TaraCharacter.h
@@ -163,7 +163,7 @@ public:
 	float CurrentSatchels = MaxSatchels;
 
 	UPROPERTY(BlueprintReadWrite, EditAnywhere, Category = "Player|Satchel")
-	int	MaxSatchels = 2;
+	float	MaxSatchels = 2.f;
